requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
fonttools==4.56.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
pytz==2025.1
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.1
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
